feat: Add option for log rotatation by local timezone
This commit updates the time dependency in the `tracing-appender` crate to version 0.3.0. It introduces a new import `UtcOffset` and modifies the `RollingFileAppender` struct to include an `offset` field of type `UtcOffset`. The `now` method in the `RollingFileAppender` implementation has also been modified to use the `offset` field when returning the current time.